Rename prefix from coroutine to suspend
