+++ coroutines/src/main/java/com/parse/coroutines/ParseUserCoroutinesExtensions.kt
@@ -1,13 +1,13 @@
 @file:JvmName("ParseUserCoroutinesExtensions")
 
-package com.parse.coroutines.user
+package com.parse.coroutines
 
 import com.parse.ParseUser
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 import kotlin.coroutines.suspendCoroutine
 
-suspend fun ParseUser.coroutineSignUp(): ParseUser {
+suspend fun ParseUser.suspendSignUp(): ParseUser {
     return suspendCoroutine { continuation ->
         signUpInBackground { e ->
             if (e == null) continuation.resume(this)
